Improvements on test structure
