--- src/tests/globalTeardown.ts
@@ -1,6 +1,0 @@
-import mongoose from 'mongoose'
-
-export default async () => {
-  await mongoose.connection.close()
-  await global.__MONGOD__.stop()
-}
